restructured page-encointer
