+++ lib/page-encointer/meetup/attestation/components/qrCode.dart
