+++ lib/page-encointer/meetup/attestation/components/scanQrCode.dart
